+++ gzap/CHANGELOG.md
@@ -5,6 +5,8 @@ This is a bugfix release.
 - New:
   - Globbing expressions are now supported in `included_levels`, `excluded_levels`,
     and `starting_levels`.
+  - `start_with_keys` option can be used to turn off starting with all keys for
+    the `starting_levels`.
 - Change:
   - Inventory is now sorted alphabetically.
   - `included_item_categories` option now supports all recognized item categories.